+++ Westwind.Utilities/Utilities/HtmlUtils.cs
@@ -151,8 +151,8 @@ namespace Westwind.Utilities
         /// <param name="imageBytes"></param>
         /// <param name="mimeType"></param>
         /// <returns></returns>
-        public static string BinaryToEmbeddedBase64(byte[] imageBytes, string mimeType = "image/png" )
-        {             
+        public static string BinaryToEmbeddedBase64(byte[] imageBytes, string mimeType = "image/png")
+        {
             var data = $"data:{mimeType};base64," + Convert.ToBase64String(imageBytes, 0, imageBytes.Length);
             return data;
         }
@@ -266,5 +266,113 @@ namespace Westwind.Utilities
 
             return html;
         }
+
+        /// <summary>
+        /// Create an Href HTML link.
+        /// 
+        /// 
+        /// <remarks>
+        /// For NetFx in ASPX applications ~ resolves to the application root
+        /// </remarks>
+        /// </summary>
+        /// <param name="text"></param>
+        /// <param name="url"></param>
+        /// <param name="target"></param>
+        /// <param name="additionalMarkup"></param>
+        /// <returns></returns>
+        public static string Href(string text, string url, string target = null, string additionalMarkup = null)
+        {
+
+#if NETFULL
+            if (url.StartsWith("~"))
+                url = ResolveUrl(url);
+#endif
+
+            return "<a href=\"" + url + "\" " +
+                   (string.IsNullOrEmpty(target) ? string.Empty : "target=\"" + target + "\" ") +
+                   (string.IsNullOrEmpty(additionalMarkup) ? string.Empty : additionalMarkup) +
+                   ">" + text + "</a>";
+        }
+
+        /// <summary>
+        /// Creates an HREF HTML Link
+        /// </summary>
+        /// <param name="url"></param>
+        public static string Href(string url)
+        {
+            return Href(url, url, null, null);
+        }
+
+        /// <summary>
+        /// Returns an IMG link as a string. If the image is null
+        /// or empty a blank string is returned.
+        /// </summary>
+        /// <param name="imageUrl"></param>
+        /// <param name="additionalMarkup">any additional attributes added to the element</param>
+        /// <returns></returns>
+        public static string ImgRef(string imageUrl, string additionalMarkup = null)
+        {
+            if (string.IsNullOrEmpty(imageUrl))
+                return string.Empty;
+
+#if NETFULL
+            if (imageUrl.StartsWith("~"))
+                imageUrl = ResolveUrl(imageUrl);
+#endif
+
+            string img = "<img src=\"" + imageUrl + "\" ";
+
+            if (!string.IsNullOrEmpty("additionalMarkup"))
+                img += additionalMarkup + " ";
+
+            img += "/>";
+            return img;
+        }
+
+#if NETFULL
+        /// <summary>
+        /// Resolves a URL based on the current HTTPContext
+        /// 
+        /// Note this method is added here internally only
+        /// to support the HREF() method and ~ expansion
+        /// on urls.
+        /// </summary>
+        /// <param name="originalUrl"></param>
+        /// <returns></returns>
+        public static string ResolveUrl(string originalUrl)
+        {
+            if (string.IsNullOrEmpty(originalUrl))
+                return string.Empty;
+
+            // Absolute path - just return
+            if (originalUrl.IndexOf("://") != -1)
+                return originalUrl;
+
+            // Fix up image path for ~ root app dir directory
+            if (originalUrl.StartsWith("~"))
+            {
+                //return VirtualPathUtility.ToAbsolute(originalUrl);
+                string newUrl = "";
+
+                // Avoid pulling in System.Web reference
+                dynamic context = ReflectionUtils.GetStaticProperty( "System.Web.HttpContext", "Current");
+                if (context != null)
+                {
+                    newUrl = context.Request.ApplicationPath +
+                          originalUrl.Substring(1);
+                    newUrl = newUrl.Replace("//", "/"); // must fix up for root path
+                }
+                else
+                    // Not context: assume current directory is the base directory
+                    throw new ArgumentException("Invalid URL: Relative Urls are only available in classic ASP.NET Web applications.");
+
+                // Just to be sure fix up any double slashes
+                return newUrl;
+            }
+
+            return originalUrl;
+        }
+#endif
+
     }
 }